pyproject.toml
[project]
name = "advent-of-code"
version = "0.1.0"
description = "Advent of Code Repo"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "lark>=1.2.2",
    "more-itertools>=10.6.0",
    "networkx>=3.4.2",
    "numpy>=2.2.4",
    "parse>=1.20.2",
    "pillow>=11.3.0",
    "pre-commit>=4.3.0",
    "pytest>=8.3.3",
    "regex>=2024.9.11",
    "ruff>=0.13.0",
    "sortedcontainers>=2.4.0",
    "sympy>=1.14.0",
    "ty>=0.0.1a20",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
